+++ config/src/color.rs
@@ -185,19 +185,6 @@ impl Default for TabBarColors {
 #[derive(Debug, Deserialize, Serialize, Clone)]
 pub struct TabBarStyle {
     #[serde(default = "default_tab_left")]
-    pub active_tab_left: String,
-    #[serde(default = "default_tab_right")]
-    pub active_tab_right: String,
-    #[serde(default = "default_tab_left")]
-    pub inactive_tab_left: String,
-    #[serde(default = "default_tab_right")]
-    pub inactive_tab_right: String,
-    #[serde(default = "default_tab_left")]
-    pub inactive_tab_hover_left: String,
-    #[serde(default = "default_tab_right")]
-    pub inactive_tab_hover_right: String,
-
-    #[serde(default = "default_tab_left")]
     pub new_tab_left: String,
     #[serde(default = "default_tab_right")]
     pub new_tab_right: String,
@@ -210,12 +197,6 @@ pub struct TabBarStyle {
 impl Default for TabBarStyle {
     fn default() -> Self {
         Self {
-            active_tab_left: default_tab_left(),
-            active_tab_right: default_tab_right(),
-            inactive_tab_left: default_tab_left(),
-            inactive_tab_right: default_tab_right(),
-            inactive_tab_hover_left: default_tab_left(),
-            inactive_tab_hover_right: default_tab_right(),
             new_tab_left: default_tab_left(),
             new_tab_right: default_tab_right(),
             new_tab_hover_left: default_tab_left(),